Properly convert from NWU to ENU in IMU messages
IMU orientations are given in a yaw-fixed frame, not a coordinate frame
firmly fixed along all three axes. The previous quaternion rotation was
rotating the entire coordinate frame, causing the sensor output to
report weird values for roll and pitch.

The end goal was to add 90 degrees to the yaw output, so that's what we
do now.