Obtain good local EKF config, semi-working global
The issue with the state estimate climbing to God was due to a bad tf
frame setting: the SONAR was reporting an altitude of, e.g. 0.5m in
`um_base_link`, which was the _local_ frame of the quadcopter.

In other words, when the EKF asked it for the vehicle's altitude, it
reported that the vehicle was "0.5m above where it currently is right
now."

Have it report in the um_odom frame, and things are working out more
nicely.